requirements.txt
PyGObject==3.42.1
pycairo==1.21.0
